:art: Move enchantment API classes
