--- src/main/java/org/sandboxpowered/sandbox/api/enchantment/Target.java
@@ -1,6 +1,0 @@
-package org.sandboxpowered.sandbox.api.enchantment;
-
-import org.sandboxpowered.sandbox.api.item.Item;
-
-public enum Target {
-}
